# 智能土地.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "土地监测系统后端接口文档。\n用户可认领土地，通过线上支付发送施肥浇水请求，管理员端接收请求人工执行。\n",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/lands": {
      "get": {
        "summary": "获取土地列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "土地"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页条数",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 50
            }
          },
          {
            "name": "zone",
            "in": "query",
            "description": "区域筛选 A/B/C",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "土地状态筛选",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "AVAILABLE",
                "CLAIMED",
                "ALL"
              ],
              "default": "ALL"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PageResultLand"
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/lands/{landId}": {
      "get": {
        "summary": "获取土地详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "土地"
        ],
        "parameters": [
          {
            "name": "landId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LandDetail"
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/lands/{landId}/claim": {
      "post": {
        "summary": "认领土地",
        "deprecated": false,
        "description": "",
        "tags": [
          "土地"
        ],
        "parameters": [
          {
            "name": "landId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "duration"
                ],
                "properties": {
                  "duration": {
                    "type": "integer",
                    "enum": [
                      1,
                      3,
                      6,
                      12
                    ],
                    "description": "认领时长（月）"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ClaimResult"
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/lands/{landId}/monitor": {
      "get": {
        "summary": "获取土地监测数据",
        "deprecated": false,
        "description": "",
        "tags": [
          "监测数据"
        ],
        "parameters": [
          {
            "name": "landId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "range",
            "in": "query",
            "description": "历史范围",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "24h",
                "7d",
                "30d"
              ],
              "default": "24h"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MonitorData"
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/monitor/report": {
      "post": {
        "summary": "上报监测数据",
        "deprecated": false,
        "description": "",
        "tags": [
          "监测数据"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "landId"
                ],
                "properties": {
                  "landId": {
                    "type": "integer"
                  },
                  "temperature": {
                    "type": "number",
                    "format": "double",
                    "description": "温度℃"
                  },
                  "humidity": {
                    "type": "number",
                    "format": "double",
                    "description": "空气湿度%"
                  },
                  "soilMoisture": {
                    "type": "number",
                    "format": "double",
                    "description": "土壤湿度%"
                  },
                  "lightIntensity": {
                    "type": "number",
                    "format": "double",
                    "description": "光照强度lux"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/admin/lands/{landId}": {
      "put": {
        "summary": "修改土地信息",
        "deprecated": false,
        "description": "",
        "tags": [
          "管理员"
        ],
        "parameters": [
          {
            "name": "landId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LandCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "summary": "获取通知列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "通知"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页条数",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 50
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "SYSTEM",
                "SERVICE",
                "PAY"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PageResultNotification"
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/notifications/{notificationId}/read": {
      "put": {
        "summary": "标记通知已读",
        "deprecated": false,
        "description": "",
        "tags": [
          "通知"
        ],
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/notifications/read-all": {
      "put": {
        "summary": "全部标记已读",
        "deprecated": false,
        "description": "",
        "tags": [
          "通知"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "0成功，其他为错误码",
            "example": 0
          },
          "msg": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        }
      },
...
      "LandItem": {
        "type": "object",
        "properties": {
          "landId": {
            "type": "integer"
          },
          "landName": {
            "type": "string",
            "example": "A区-01号田"
          },
          "zone": {
            "type": "string",
            "example": "A"
          },
          "area": {
            "type": "number",
            "format": "double",
            "example": 10.5
          },
          "areaUnit": {
            "type": "string",
            "example": "㎡"
          },
          "image": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "CLAIMED"
            ]
          },
          "statusName": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double",
            "example": 299
          },
          "priceUnit": {
            "type": "string",
            "example": "元/年"
          },
          "soilType": {
            "type": "string",
            "example": "黑土"
          },
          "cropType": {
            "type": "string",
            "example": "水稻"
          },
          "claimUser": {
            "type": "string",
            "nullable": true
          },
          "claimExpireTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "LandDetail": {
        "type": "object",
        "properties": {
          "landId": {
            "type": "integer"
          },
          "landName": {
            "type": "string"
          },
          "zone": {
            "type": "string"
          },
          "area": {
            "type": "number",
            "format": "double"
          },
          "areaUnit": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string"
          },
          "statusName": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "priceUnit": {
            "type": "string"
          },
          "soilType": {
            "type": "string"
          },
          "cropType": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "claimUser": {
            "type": "object",
            "properties": {
              "userId": {
                "type": "integer"
              },
              "nickname": {
                "type": "string"
              },
              "avatar": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "claimTime": {
            "type": "string",
            "format": "date-time"
          },
          "claimExpireTime": {
            "type": "string",
            "format": "date-time"
          },
          "latestMonitor": {
            "$ref": "#/components/schemas/MonitorLatest"
          },
          "servicePrice": {
            "type": "object",
            "properties": {
              "waterPrice": {
                "type": "number",
                "format": "double",
                "example": 15
              },
              "fertilizePrice": {
                "type": "number",
                "format": "double",
                "example": 35
              },
              "priceUnit": {
                "type": "string",
                "example": "元/次"
              }
            }
          }
        }
      },
      "LandCreateRequest": {
        "type": "object",
        "required": [
          "landName",
          "zone",
          "area",
          "price",
          "waterPrice",
          "fertilizePrice"
        ],
        "properties": {
          "landName": {
            "type": "string"
          },
          "zone": {
            "type": "string"
          },
          "area": {
            "type": "number",
            "format": "double"
          },
          "areaUnit": {
            "type": "string",
            "default": "㎡"
          },
          "soilType": {
            "type": "string"
          },
          "cropType": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "认领价格"
          },
          "waterPrice": {
            "type": "number",
            "format": "double",
            "description": "浇水价格"
          },
          "fertilizePrice": {
            "type": "number",
            "format": "double",
            "description": "施肥价格"
          },
          "description": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ClaimResult": {
        "type": "object",
        "properties": {
          "claimId": {
            "type": "integer"
          },
          "landId": {
            "type": "integer"
          },
          "landName": {
            "type": "string"
          },
          "duration": {
            "type": "integer",
            "description": "认领时长（月）"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "paymentNo": {
            "type": "string",
            "description": "支付单号"
          },
          "status": {
            "type": "string",
            "example": "PENDING_PAY"
          },
          "expireTime": {
            "type": "string",
            "format": "date-time",
            "description": "支付过期时间"
          }
        }
      },
...
      "MonitorLatest": {
        "type": "object",
        "properties": {
          "temperature": {
            "type": "number",
            "format": "double",
            "description": "温度℃"
          },
          "humidity": {
            "type": "number",
            "format": "double",
            "description": "空气湿度%"
          },
          "soilMoisture": {
            "type": "number",
            "format": "double",
            "description": "土壤湿度%"
          },
          "lightIntensity": {
            "type": "number",
            "format": "double",
            "description": "光照强度lux"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MonitorData": {
        "type": "object",
        "properties": {
          "landId": {
            "type": "integer"
          },
          "landName": {
            "type": "string"
          },
          "latest": {
            "$ref": "#/components/schemas/MonitorLatest"
          },
          "history": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "time": {
                  "type": "string",
                  "format": "date-time"
                },
                "temperature": {
                  "type": "number",
                  "format": "double"
                },
                "humidity": {
                  "type": "number",
                  "format": "double"
                },
                "soilMoisture": {
                  "type": "number",
                  "format": "double"
                },
                "lightIntensity": {
                  "type": "number",
                  "format": "double"
                }
              }
            }
          }
        }
      },
      "NotificationItem": {
        "type": "object",
        "properties": {
          "notificationId": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "SYSTEM",
              "SERVICE",
              "PAY"
            ]
          },
          "isRead": {
            "type": "boolean"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "extra": {
            "type": "object",
            "properties": {},
            "nullable": true
          }
        }
      },
...
      "PageResultLand": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LandItem"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
...
      "PageResultNotification": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationItem"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "unreadCount": {
            "type": "integer"
          }
        }
      }
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}